This flow is only used in the archivist, so move it there
