+++ archivist/src/main/scala/uk/ac/wellcome/platform/archive/archivist/flow/ArchiveDigestItemJobFlow.scala
@@ -5,10 +5,7 @@ import akka.stream.scaladsl.Flow
 import com.amazonaws.services.s3.AmazonS3
 import grizzled.slf4j.Logging
 import uk.ac.wellcome.platform.archive.archivist.models.ArchiveDigestItemJob
-import uk.ac.wellcome.platform.archive.common.flows.{
-  FoldEitherFlow,
-  OnErrorFlow
-}
+import uk.ac.wellcome.platform.archive.common.flows.FoldEitherFlow
 import uk.ac.wellcome.platform.archive.common.models.error.ArchiveError
 
 object ArchiveDigestItemJobFlow extends Logging {
