+++ archivist/src/main/scala/uk/ac/wellcome/platform/archive/archivist/flow/OnErrorFlow.scala
@@ -1,4 +1,5 @@
-package uk.ac.wellcome.platform.archive.common.flows
+package uk.ac.wellcome.platform.archive.archivist.flow
+
 import akka.stream.scaladsl.Flow
 import grizzled.slf4j.Logging
 import uk.ac.wellcome.platform.archive.common.models.error.ArchiveError
