+++ archivist/src/main/scala/uk/ac/wellcome/platform/archive/archivist/flow/UploadItemFlow.scala
@@ -11,10 +11,7 @@ import uk.ac.wellcome.platform.archive.archivist.models.{
   ZipLocation
 }
 import uk.ac.wellcome.platform.archive.archivist.zipfile.ZipFileReader
-import uk.ac.wellcome.platform.archive.common.flows.{
-  FoldEitherFlow,
-  OnErrorFlow
-}
+import uk.ac.wellcome.platform.archive.common.flows.FoldEitherFlow
 import uk.ac.wellcome.platform.archive.common.models.error.ArchiveError
 
 /** This flow extracts an item from a ZIP file, uploads it to S3 and calculates the checksum
